requirements.txt
pynput
cryptography
pillow
jinja2
pywin32